Move msg to common crate
