+++ common/cosmwasm-smart-contracts/service-provider-directory/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod types;
+pub mod msg;
 
 // Re-export all types at the top-level
 pub use types::*;
