+++ contracts/service-provider-directory/src/integration_tests.rs
@@ -1,12 +1,14 @@
 //! Integration tests using cw-multi-test.
 
 use cosmwasm_std::Addr;
-use nym_service_provider_directory_common::{NymAddress, Service, ServiceType};
+use nym_service_provider_directory_common::{
+    msg::{ConfigResponse, PagedServicesListResponse, ServiceInfo},
+    NymAddress, Service, ServiceType,
+};
 
 use crate::{
     constants::SERVICE_DEFAULT_RETRIEVAL_LIMIT,
     error::ContractError,
-    msg::{ConfigResponse, PagedServicesListResponse, ServiceInfo},
     test_helpers::{helpers::nyms, test_setup::TestSetup},
 };
