+++ contracts/service-provider-directory/src/state/config.rs
@@ -1,5 +1,6 @@
 use cosmwasm_std::{Coin, Storage};
 use cw_storage_plus::Item;
+use nym_service_provider_directory_common::msg::ConfigResponse;
 use serde::{Deserialize, Serialize};
 
 use crate::error::Result;
@@ -10,6 +11,14 @@ const CONFIG: Item<Config> = Item::new(CONFIG_KEY);
 #[derive(Serialize, Deserialize, Clone, PartialEq, Debug)]
 pub(crate) struct Config {
     pub deposit_required: Coin,
+}
+
+impl From<Config> for ConfigResponse {
+    fn from(config: Config) -> Self {
+        ConfigResponse {
+            deposit_required: config.deposit_required,
+        }
+    }
 }
 
 pub(crate) fn save_config(store: &mut dyn Storage, config: &Config) -> Result<()> {
