+++ contracts/service-provider-directory/src/test_helpers/assert.rs
@@ -1,11 +1,10 @@
 use cosmwasm_std::{from_binary, testing::mock_env, Addr, Coin, Deps, StdError};
-use nym_service_provider_directory_common::ServiceId;
-
-use crate::{
-    constants::SERVICE_DEFAULT_RETRIEVAL_LIMIT,
-    error::ContractError,
+use nym_service_provider_directory_common::{
     msg::{ConfigResponse, PagedServicesListResponse, QueryMsg, ServiceInfo},
+    ServiceId,
 };
+
+use crate::{constants::SERVICE_DEFAULT_RETRIEVAL_LIMIT, error::ContractError};
 
 pub fn assert_config(deps: Deps, admin: &Addr, deposit_required: Coin) {
     crate::state::assert_admin(deps, admin).unwrap();
