+++ contracts/service-provider-directory/src/test_helpers/helpers.rs
@@ -1,8 +1,5 @@
 use cosmwasm_std::{Coin, Response};
 use cw_multi_test::AppResponse;
-use nym_service_provider_directory_common::Service;
-
-use crate::msg::ExecuteMsg;
 
 pub fn nyms(amount: u64) -> Coin {
     Coin::new(amount.into(), "unym")
@@ -27,11 +24,3 @@ pub fn get_app_attribute(response: &AppResponse, key: &str) -> String {
         .clone()
 }
 
-impl From<Service> for ExecuteMsg {
-    fn from(service: Service) -> Self {
-        ExecuteMsg::Announce {
-            nym_address: service.nym_address,
-            service_type: service.service_type,
-        }
-    }
-}
