+++ contracts/service-provider-directory/src/test_helpers/test_setup.rs
@@ -1,16 +1,16 @@
 use anyhow::Result;
 use cosmwasm_std::{coins, Addr, Coin, StdResult, Uint128};
 use cw_multi_test::{App, AppBuilder, AppResponse, ContractWrapper, Executor};
-use nym_service_provider_directory_common::{NymAddress, ServiceId, ServiceType};
-use serde::de::DeserializeOwned;
-
-use crate::{
+use nym_service_provider_directory_common::{
     msg::{
         ConfigResponse, ExecuteMsg, InstantiateMsg, PagedServicesListResponse, QueryMsg,
         ServiceInfo,
     },
-    test_helpers::helpers::get_app_attribute,
+    NymAddress, ServiceId, ServiceType,
 };
+use serde::de::DeserializeOwned;
+
+use crate::test_helpers::helpers::get_app_attribute;
 
 const DENOM: &str = "unym";
 const ADDRESSES: &[&str] = &[
